+++ packages/ui/postcss.config.mjs
@@ -1,11 +1,6 @@
 import postcssCustomMedia from 'postcss-custom-media';
-import postcssCsso from 'postcss-csso';
 import postcssImport from 'postcss-import';
 
 export default {
-  plugins: [
-    postcssImport(),
-    postcssCustomMedia({ preserve: false }),
-    postcssCsso(),
-  ],
+  plugins: [postcssImport(), postcssCustomMedia({ preserve: false })],
 };